+++ src/store/index.ts
@@ -77,6 +77,7 @@ export interface PackageInfoStore {
     latest: string | null;
     manager: string;
     is_outdated: boolean;
+    update_checked?: boolean;
     description: string | null;
 }